feat: 소수점 거래 추가
- BuyRequest, SellRequest quantity: Integer → Float 변경
- GroupBuyRequest, GroupSellRequest quantity: Integer → Float 변경
- 최소 수량 검증: 0.01 이상으로 변경 (@DecimalMin)
- InsufficientHoldingException: int → float 파라미터 지원
- 그룹 분할 매매에서 소수점 단위 거래 가능 (예: 4명이 1주 매수 시 각자 0.25주)

+++ trading-service/src/main/java/com/example/trading_service/exception/InsufficientHoldingException.java
@@ -3,7 +3,7 @@ package com.example.trading_service.exception;
 import java.util.Map;
 
 public class InsufficientHoldingException extends BusinessException {
-    public InsufficientHoldingException(int required, int available) {
+    public InsufficientHoldingException(float required, float available) {
         super("보유 수량이 부족합니다", "INSUFFICIENT_HOLDING", 
               Map.of("required", required, "available", available));
     }
